requirements.txt
feedparser
python-vlc
tinydb